Se esta trabajando en la vista de ordenes de trabajo y se modifico la vista del cliente

+++ lib/features/ordenes_servicio/orden_servicio/data/models/items_servicio/item_con_precio_model/item_con_precio_model.dart
@@ -10,7 +10,7 @@ sealed class ItemConPrecioModel with _$ItemConPrecioModel {
     required int id,
     required int ordenId,
     required PrecioConDetallesModel precio,
-    required int cantidad,
+    required double cantidad,
     required double importe,
     required bool estaEntregado,
     required DateTime fechaEntrega,
